maze solved
C++ homework 3-4 nobody left so change children from int to double,
solved;
Maze solved, but I don’t know why use the pointer failed. Code so ugly.

+++ 计算机程序设计（C++）西安交通大学/Week3/3-4搬砖问题.cpp
@@ -18,7 +18,7 @@ int main(int argc, const char * argv[]) {
 		for (women=0;women<=max-men;women++)
 		{
 			children=max-men-women;
-			if (4*men+3*women+children/2==max)
+			if (4*men+3*women+(double)children/2==max)
 			{
 				print_result(men,women,children);
 				possible=false;
